+++ lib/src/utils/analyzer_utils.dart
@@ -1,10 +1,14 @@
 // ignore_for_file: implementation_imports
+import 'package:analyzer/dart/analysis/analysis_context.dart';
 import 'package:analyzer/dart/analysis/analysis_context_collection.dart';
 import 'package:analyzer/file_system/physical_file_system.dart';
 import 'package:analyzer/src/dart/analysis/analysis_context_collection.dart';
 import 'package:analyzer/src/dart/analysis/byte_store.dart';
 import 'package:analyzer/src/dart/analysis/file_byte_store.dart';
+import 'package:glob/glob.dart';
 import 'package:path/path.dart';
+
+import 'file_utils.dart';
 
 AnalysisContextCollection createAnalysisContextCollection(
   Iterable<String> folders,
@@ -20,6 +24,22 @@ AnalysisContextCollection createAnalysisContextCollection(
     resourceProvider: resourceProvider,
     byteStore: _createByteStore(resourceProvider),
   );
+}
+
+Set<String> getFilePaths(
+  Iterable<String> folders,
+  AnalysisContext context,
+  String rootFolder,
+  Iterable<Glob> excludes,
+) {
+  final contextFolders = folders.where((path) {
+    final newPath = normalize(join(rootFolder, path));
+
+    return newPath == context.contextRoot.root.path ||
+        context.contextRoot.root.path.startsWith('$newPath/');
+  }).toList();
+
+  return extractDartFilesFromFolders(contextFolders, rootFolder, excludes);
 }
 
 /// If the state location can be accessed, return the file byte store,